fix: resolve bugs isNumber function and routes update delete TypeDisabled

+++ src/controllers/TypesDisableControllers/typesDisabledDeleteControllers.js
@@ -1,9 +1,23 @@
 const typesDisabledDeleteService = require("../../services/TypesDisabledServices/typesDisabledDeleteServices");
+const nameExists = require('../../services/TypesDisabledServices/nameExists');
 
 const typeDisabledDeleteController = {
   delete: async (req, res) => {
     try {
       const { name } = req.params;
+
+      const nameValid = await nameExists.validNameExists(name)
+
+      if (!nameValid) {
+        const data = {
+          code: 400,
+          erro: "Tipo de deficiência não encontrada!",
+          message: "Busque um tipo de deficiência pelo nome correto.",
+          success: false,
+        }
+
+        return res.status(400).json(data)
+      }
 
       const type = await typesDisabledDeleteService.delete(name);
